# EndPointGo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la gestion de reservas de cine",
        "title": "API Cine",
        "contact": {},
        "version": "1.0"
    },
    "host": "cinebd.ddns.net:1320",
    "basePath": "/v1",
    "paths": {
        "/cartelera": {
            "get": {
                "description": "Devuelve la cartelera del cine para un periodo de tiempo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cartelera"
                ],
                "summary": "Cartelera para un periodo de tiempo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha inicio (YYYY-MM-DD)",
                        "name": "fechaInicio",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha fin (YYYY-MM-DD)",
                        "name": "fechaFin",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Cartelera"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ocupacion/reporte": {
            "get": {
                "description": "Devuelve las ventas del cine para una pelicula en un periodo de tiempo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reportes"
                ],
                "summary": "Reporte de ocupacion por pelicula",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la pelicula",
                        "name": "idPelicula",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha inicio (YYYY-MM-DD)",
                        "name": "fechaInicio",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha fin (YYYY-MM-DD)",
                        "name": "fechaFin",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ReporteOcupacion"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/precio/{idFuncion}": {
            "get": {
                "description": "Calcula el precio final segun genero y sala",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "funciones"
                ],
                "summary": "Determina el precio de una entrada",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la funcion",
                        "name": "idFuncion",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PrecioEntrada"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservas": {
            "post": {
                "description": "Crea una reserva para una funcion con un DNI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Reserva una butaca",
                "parameters": [
                    {
                        "description": "Datos de la reserva",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Reserva"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Cartelera": {
            "type": "object",
            "properties": {
                "duracion": {
                    "type": "string"
                },
                "entradasDisponibles": {
                    "type": "integer"
                },
                "fechaHora": {
                    "type": "string"
                },
                "genero": {
                    "type": "string"
                },
                "idFuncion": {
                    "type": "integer"
                },
                "idSala": {
                    "type": "number"
                },
                "pelicula": {
                    "type": "string"
                },
                "precio": {
                    "type": "number"
                },
                "sala": {
                    "type": "string"
                }
            }
        },
        "models.PrecioEntrada": {
            "type": "object",
            "properties": {
                "precio": {
                    "type": "number"
                }
            }
        },
        "models.ReporteOcupacion": {
            "type": "object",
            "properties": {
                "butacasVendidas": {
                    "type": "integer"
                },
                "fechaInicio": {
                    "type": "string"
                },
                "idFuncion": {
                    "type": "integer"
                },
                "idSala": {
                    "type": "integer"
                },
                "recaudacion": {
                    "type": "number"
                },
                "sala": {
                    "type": "string"
                }
            }
        },
        "models.Reserva": {
            "type": "object",
            "properties": {
                "dni": {
                    "type": "string"
                },
                "idButaca": {
                    "type": "integer"
                },
                "idFuncion": {
                    "type": "integer"
                }
            }
        }
    }
}